Logout implemented
Added a "Log Out" option to the options menu. Clicking it should take
the user back to the login activity.

Each Activity class (excluding Logger and RegisterUser) now has a
logOut() function which is called when either the Log Out option is
selected or  a later Activity has called logOut(). logOut() causes an
Activity and all preceding activites to call finish() down to the Logger
activity.